Refactored GroupDetails container, added ModalPopup component, added ItemWithButtonList component
