+++ src/components/shared/ItemWithButtonList.test.js
@@ -1,0 +1,23 @@
+import React from 'react'
+import ItemWithButtonList from './ItemWithButtonList'
+import renderer from 'react-test-renderer'
+
+describe('Component: ListItemWithButton', () => {
+
+    it('renders a snapshot', () => {
+
+        const itemList = [
+            {id: 1, name: 'ok'},
+            {id: 2, name: 'pvp'},
+            {id: 3, name: 'lol'},
+        ]
+
+        const tree = renderer.create(
+            <ItemWithButtonList
+                itemList={itemList}
+                buttonText="ok"
+            />
+        ).toJSON()
+        expect(tree).toMatchSnapshot()
+    })
+})
